requirements.txt
PyYAML
werkzeug
flask
Pillow
Crypto